requirements.txt
Django>=3.1
djangorestframework>=3.11.1
psycopg2>=2.8.5
Pillow>=5.4.0

flake8>=3.8.3
